+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+# edfi_api_client v0.2.0
+## New Features
+- `EdFiClient.get_swagger()` now returns an EdFiSwagger class that parses OpenAPI Swagger specification.
+- `EdFiClient.resources` and `EdFiClient.descriptors` lazily retrieves lists of respective endpoints from Swagger.
+- `EdFiEndpoint` child class attributes `description` and `has_deletes` lazily retrieves this metadata from Swagger.
+
+## Under the hood
+- Requests re-authenticate automatically, based on the expiration-time retrieved from the API.
+
+
 # edfi_api_client v0.1.4
 ## Fixes
 - Compatibility fix for Ed-Fi 6.0: casing changed for change version API responses
+
 
 # edfi_api_client v0.1.2
 ## New features
@@ -12,9 +23,11 @@
 ## Fixes
 - Fix bug in `EdFiResource.get_pages()` where default `change_version_step_size` was used instead of argument
 
+
 # edfi_api_client v0.1.1
 ## Fixes
-- retry on 500 errors
+- Retry on 500 errors
+
 
 # edfi_api_client v0.1.0
 Initial release